checkpoint: verified Death-B fix + STATE-12 + WIP tree
bt/store.c (new): flash layer baseline + Death-B fix (drop outer flash_safe_execute; SDK irq_state[] is not nest-safe) + store_irq_guard + BCON pm (main.c)

Verified 2026-09-16: Death-B gone (SUB complete, pm=0), op-in negative in both flavors, host 3/3 green, FW builds

As-is (not HW-verified): STATE-12 sticks, Phase-3 wireless, Plan-A partial, docs/wiki/history/ledger

+++ src/usb/usb_hid.h
@@ -14,10 +14,11 @@ int usb_build_81_reply(const uint8_t *req, int req_len, uint8_t *out,
 bool usb_req_is_handshake(const uint8_t *req, int req_len);
 
 /* 0x01 サブコマンド応答 (64B の 0x21 レポート) の入力文脈。
- * btn は pack済み 3B、stick は PC 側 8bit 値、mac は自アドレス。 */
+ * btn は pack済み 3B、stick は 12bit 値（中央 0x800。LEN=8受信は取込時に<<4済み）、
+ * mac は自アドレス。 */
 typedef struct {
     uint8_t btn[3];
-    uint8_t lx, ly, rx, ry;
+    uint16_t lx, ly, rx, ry;
     uint8_t timer;
     uint8_t mac[6];
     uint8_t player;
